Split the web/sharings/sharings.go file
